requirements.txt
python-decouple
django-cors-headers
mysqlclient